+++ README.md
@@ -52,42 +52,30 @@ X videos with status `new` remaining
 
 ## Запуск проекта
 
-1. Для автоматического получения новых лайкнутых видео по юзернейму вам потребуется `msToken` (будь он проклят).
-
-   1.1. Достать его можно из cookies браузера после захода в TikTok
-   
-   1.2. Или с помощью скрипта `python ms_token/get_token.py`
-
-   Главная боль в том, что у меня никак не получается заставить этот скрипт работать внутри докера и **отдавать рабочий токен**. Да, токен генерируется, но для получения видео не годится. В директории [ms_token/](https://github.com/skrepkaq/ATP/tree/ms-token/ms_token) (отдельная ветка ms-token) есть несколько моих экспериментов на этот счёт, можете попробовать довести их до ума.
-   [Подробнее вот тут](https://github.com/skrepkaq/ATP/blob/ms-token/ms_token/README.md).
-   Так вот:
-
-   1.3. Ставим TikTokApi `pip install TikTokApi`
-   
-   1.4. Playwright браузер `playwright install chromium`
-
-   1.5  Запускаем `python ms_token/get_token.py`
-
-   1.3. Полученный токен вставляем в `.env`
-
-   1.4. Убедитесь что ваши лайкнутые видео видно всем (в настройках приватности)
-
-2. Заполните остальные параметры в `.env`:
+1. Заполните параметры в `.env`:
    - `TIKTOK_USER` - ваш username в TikTok
    - `TELEGRAM_BOT_TOKEN` - токен вашего бота (получите у @BotFather)
    - `TELEGRAM_CHAT_ID` - ID чата/канала для отправки уведомлений
    - `TZ` - ваш часовой пояс (например, Europe/Moscow)
 
-#### Шаг 1 можно пропустить и поставить `DOWNLOAD_FROM_TIKTOK=false` в `.env`.
+2. Убедитесь что ваши лайкнутые видео видно всем (в настройках приватности)
+
+#### Вы можете поставить `DOWNLOAD_FROM_TIKTOK=false` в `.env`.
 #### Тогда видео всё ещё будут проверяться на доступность, но не будут автоматически импортироваться из TikTok
 
-### Docker
+### Docker (рекомендуется)
 
 ```bash
 docker compose up -d
 ```
 
-### Без docker
+### Запуск без Docker
+
+Как-то запустите browserless/chromium. Можно в докере, только его:
+
+1. Раскомментируйте ports в compose.yaml
+2. ```docker compose up -d browserless```
+3. Запустите проект
 
 ```bash
 python entrypoint.py
@@ -103,14 +91,9 @@ python entrypoint.py
 
 ## Устранение неполадок
 
-### Проблемы с msToken
-Если вы видите сообщение "msToken is invalid":
-1. Получите новый токен через скрипт или из браузера
-2. Обновите токен в `.env`
-3. Перезапустите приложение
-
 ### Проблемы с доступом к TikTok
 Если видео не скачиваются:
+
 0. Удалите БД если только что создали её, что бы в ней не было failed видео
 1. Проверьте доступность TikTok из вашей сети
 2. При необходимости настройте VPN
