+++ example.env
@@ -7,13 +7,15 @@ DOWNLOADS_DIR=./downloads
 TIKTOK_DATA_FILE=./user_data_tiktok.json
 
 # Настройки Telegram
-TELEGRAM_BOT_TOKEN="your_telegram_bot_token"
-TELEGRAM_CHAT_ID="your_telegram_chat_id"
+TELEGRAM_BOT_TOKEN=""
+TELEGRAM_CHAT_ID=""
 
 # Настройки тиктока
 DOWNLOAD_FROM_TIKTOK=true
-MS_TOKEN=""
-TIKTOK_USER="user"
+TIKTOK_USER=""
+
+# Настройки browserless
+BROWSERLESS_URL="ws://localhost:3000"
 
 # Настройки проверки доступности
 CHECK_INTERVAL_DAYS=7
