--- ms_token/POC/Dockerfile
@@ -1,7 +1,0 @@
-FROM python:3.12-slim
-
-RUN pip install --no-cache-dir TikTokApi
-
-COPY get_token.py .
-
-ENTRYPOINT ["python", "-u", "get_token.py"]
