--- ms_token/README.md
@@ -1,126 +1,0 @@
-# Эксперименты по получению msToken
-
-В этой директории собраны мои эксперименты по автоматическому получению msToken от TikTok. Этот токен необходим для работы ATP, чтобы автоматически импортировать новые лайкнутые видео.
-
-## Проблема
-
-TikTok использует систему защиты от ботов
-
-Главная сложность заключается в том, что даже если удаётся получить msToken, он часто оказывается невалидным для реального использования в API.
-
-
-## Тестирование
-
-Для проверки валидности полученного токена используется функция `test_token()`. Токен считается валидным только если с его помощью удаётся получить реальные данные через TikTok API.
-
-## Текущие эксперименты
-
-### Xvfb (`xvfb/`)
-Попытка использовать "настоящий" браузер через виртуальный X-сервер в Docker.
-
-**Как работает:**
-- Запускает Xvfb для создания виртуального дисплея
-- Использует реальный Chrome вместо WebDriver
-- Пытается обойти детекцию автоматизации
-
-**Настройка и запуск:**
-
-##### Docker
-
-```bash
-# В директории xvfb/
-docker compose up
-```
-
-Результат:
-
-```
-ms-token-xvfb  | Generated token:
-ms-token-xvfb  | VW4XBFb1CksZcNN3.......
-ms-token-xvfb  | Checking token...
-ms-token-xvfb  | Token is bad(
-ms-token-xvfb  | None -> TikTok returned an empty response. They are detecting you're a bot, try some of these: headless=False, browser='webkit', consider using a proxy
-```
-
-##### Без docker
-
-
-```bash
-python get_token_xvfb.py
-```
-
-Результат:
-
-```
-Generated token:
-Hyxqy8GrkReYL3bm.......
-Checking token...
-Token is ok!
-```
-
-**Статус:**
-- ✅ Работает без докера и с headless=False
-- ❌ В Docker контейнере токены не проходят валидацию
-- ❌ Не работает headless режиме
-
-
-### Selenium (`selenium/`)
-Эксперимент с использованием Selenium Grid для более реалистичной эмуляции браузера.
-
-**Как работает:**
-- Поднимает Selenium Grid с настроенным Chrome
-- Использует WebDriver для автоматизации
-- Добавляет эмуляцию человеческого поведения:
-  - Случайные задержки
-  - Движения мыши
-  - Скроллинг страницы
-
-**Настройка и запуск:**
-```bash
-# В директории selenium/
-docker compose up -d
-pip install -r requirements.txt
-python get_token_selenium.py
-```
-
-**Статус:**
-- ❌ Не работает без докера
-- ❌ В Docker тоже не работает
-
-**Основные проблемы:**
-- Сложность настройки реалистичного окружения
-- TikTok блокирует запросы с Selenium Grid
-
-## Общие проблемы всех подходов
-
-- TikTok имеет продвинутые механизмы детекции ботов
-- Не работает headless режим
-- Даже при успешном получении токена, он может быть помечен как невалидный
-- Отличия в поведении браузера в контейнере
-- Сложность эмуляции реального окружения
-
-## Возможные улучшения
-
-1. **Эмуляция поведения**
-   - Добавление случайных задержек
-   - Имитация движений мыши
-   - Естественная навигация по странице
-   - Эмуляция клавиатурного ввода
-
-2. **Работа с окружением**
-   - Настройка реалистичных user-agent
-   - Управление fingerprint браузера
-   - Сохранение и переиспользование сессий
-   - Эмуляция различных устройств
-
-3. **Обход защиты**
-   - Использование коммерческих прокси
-   - Ротация IP адресов
-   - Работа с капчей
-   - Использование реальных профилей браузера
-
-## Главная цель
-
-Любой ценой (но бесплатно) добиться генерации токена который пройдёт проверку `test_token()` (всё должно работать в Docker)
-
-Если в вас получится, создавайте pull request с вашими изменениями и забирайте **bounty** (подробнее в tg @skrepkat)
