--- ms_token/selenium/requirements.txt
@@ -1,3 +1,0 @@
-selenium==4.9.0
-webdriver-manager==3.8.5
-TikTokApi==7.1.0
